review: CountModel reads a Long counter, and the membership mutation record matches what re-running it produces
Two defects found by the task review of computenet-4ru.5.1:

1. CountModel returned Set.size (an Int) while the kernel CountCell it is
   paired with emits Propagate<CounterDelta> whose amount is a Long. ModelState
   is compared for structural equality, so ScalarState(2) != ScalarState(2L)
   and every count comparison the differential runner (computenet-4ru.8) ever
   made would have mismatched. It was also inconsistent with the model's own
   other Scalar producers, CounterSourceModel/PnCounterSourceModel, whose
   netTotal is already a Long. Reading is now a Long, the reason is in
   CountModel's KDoc beside the claim, and the test asserts the Long reading
   plus shouldNotBe the Int one.

2. MembershipSemanticsTest's class KDoc recorded the observes-always-true
   mutation as reddening 'BS-2 and the two an unobserved remove cases'.
   Re-running it (--rerun --no-build-cache, 10 tests) reddens BS-2,
   'one uncovered add among several keeps the element live' and 'an observe
   does not reach an add issued after it'; 'an unobserved remove of an element
   nobody ever added is a no-op' passes under BOTH mutations, because with no
   add on the log there is nothing for either half of the rule to cover. The
   record now says what re-running it says, with the commands and the assertion
   messages.

:oracle:test 8 files, 60 tests, 0 failures, 0 errors, 0 skipped.

+++ oracle/src/main/kotlin/civictech/oracle/model/SetOperatorModels.kt
@@ -168,10 +168,20 @@ class FlatMapSetModel(private val transform: ElementExpansion) : OperatorModel, 
  * "emits only when membership size changes" is an emission rule the batch reference cannot
  * and need not express: at quiescence the observable is the count, and effective-only
  * emission is what makes the kernel's count equal it.
+ *
+ * The count is a **`Long`**, not the `Int` `Set.size` hands back. [ModelState] is compared for
+ * *structural* equality against what a differential run folds out of the kernel terminal, and
+ * this cell's outlet is `Propagate<CounterDelta>` whose `amount` is a `Long`
+ * (kernel/src/main/kotlin/civictech/cell/data/delta/CounterDelta.kt), so a terminal over it
+ * reads `2L`, and `ScalarState(2) != ScalarState(2L)` — an `Int` here would mismatch on every
+ * case the runner (computenet-4ru.8) ever built. It is also what keeps the model's own
+ * `Scalar` family self-consistent with [CounterSourceModel]/[PnCounterSourceModel], whose
+ * `netTotal` is a `Long`. [PresenceCountModel]'s counts stay `Int` for exactly the same
+ * reason in the other direction: `PresenceCountCell`'s outlet is `MapDelta<E, Int>`.
  */
 object CountModel : OperatorModel, Serializable {
     override fun evaluate(inputs: List<ModelState>): ModelState =
-        ModelState.ScalarState(single(inputs, "CountModel").asSet("CountModel").elements.size)
+        ModelState.ScalarState(single(inputs, "CountModel").asSet("CountModel").elements.size.toLong())
 
     override fun toString(): String = "CountModel"
 }
